requirements.txt
# Core
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3


# Database
psycopg2-binary==2.9.9

# Environment
django-environ==0.11.2

# CORS
django-cors-headers==4.4.0

# API Docs
drf-spectacular==0.27.2

# QR Code
qrcode[pil]==7.4.2
Pillow==10.4.0

# Channels (WebSocket)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Production server
gunicorn==22.0.0

# Caching / Async
redis==5.0.8
celery==5.4.0

# Code Quality
black==24.8.0
flake8==7.1.1
isort==5.13.2

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
coverage==7.6.1
